build: separate out etl (dev) dependencies
