+++ README.md
@@ -1,6 +1,14 @@
 # Gene Normalization
 Services and guidelines for normalizing gene terms
 
+Installing with pip:
+
+```commandline
+pip install gene[dev]
+```
+
+The `[dev]` argument tells pip to install packages to fulfill the dependencies of the `gene.etl` package.
+
 ## Developer instructions
 Following are sections include instructions specifically for developers.
 
@@ -14,16 +22,29 @@ Once installed, from the project root dir, just run:
 ```commandline
 pipenv shell
 pipenv lock && pipenv sync
+pipenv install --dev
 ```
 
-Gene Normalization relies on [SeqRepo](https://github.com/biocommons/biocommons.seqrepo) data.
+Gene Normalization relies on [SeqRepo](https://github.com/biocommons/biocommons.seqrepo) data, which you must download yourself.
 
 From the _root_ directory:
 ```
 pip install seqrepo
 sudo mkdir /usr/local/share/seqrepo
 sudo chown $USER /usr/local/share/seqrepo
-seqrepo pull -i 2021-01-29
+seqrepo pull -i 2021-01-29  # Replace with latest version using `seqrepo list-remote-instances` if outdated
+```
+
+If you get an error similar to the one below:
+```
+PermissionError: [Error 13] Permission denied: '/usr/local/share/seqrepo/2021-01-29._fkuefgd' -> '/usr/local/share/seqrepo/2021-01-29'
+```
+
+You will want to do the following:\
+(*Might not be ._fkuefgd, so replace with your error message path*)
+```console
+sudo mv /usr/local/share/seqrepo/2021-01-29._fkuefgd /usr/local/share/seqrepo/2021-01-29
+exit
 ```
 
 ### Deploying DynamoDB Locally
