+++ gene/version.py
@@ -1,2 +1,2 @@
 """Gene normalizer version"""
-__version__ = "0.1.27"
+__version__ = "0.1.28"
