+++ requirements.txt
@@ -1,93 +1,19 @@
-#
-# These requirements were autogenerated by pipenv
-# To regenerate from the project's Pipfile, run:
-#
-#    pipenv lock --requirements
-#
-
 -i https://pypi.org/simple
-anyio==3.5.0; python_full_version >= '3.6.2'
-appdirs==1.4.4
-appnope==0.1.2; sys_platform == 'darwin'
-argcomplete==2.0.0; python_version >= '3.6'
-argh==0.26.2
-asgiref==3.5.0; python_version >= '3.7'
-asttokens==2.0.5
-attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
-backcall==0.2.0
-beautifulsoup4==4.10.0
-biocommons.seqrepo==0.6.5
-bioutils==0.5.5; python_version >= '3.6'
-boto3==1.21.26
-botocore==1.24.26; python_version >= '3.6'
-bs4==0.0.1
-canonicaljson==1.6.0; python_version ~= '3.7'
-certifi==2021.10.8
-charset-normalizer==2.0.12; python_version >= '3'
-click==8.0.4
-coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
-configparser==5.2.0; python_version >= '3.6'
-cssselect==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
-decorator==5.1.1; python_version >= '3.5'
-executing==0.8.3
-fake-useragent==0.1.11
-fastapi==0.75.0
-ga4gh.vrs[extras]==0.8.0.dev0
-ga4gh.vrsatile.pydantic==0.0.10
-gffutils==0.10.1
-h11==0.13.0; python_version >= '3.6'
-hgvs==1.5.2
-humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
-idna==3.3; python_version >= '3'
-importlib-metadata==4.11.3; python_version < '3.10'
-inflection==0.5.1; python_version >= '3.5'
-ipython==8.1.1; python_version >= '3.8'
-jedi==0.18.1; python_version >= '3.6'
-jmespath==1.0.0; python_version >= '3.7'
-jsonschema==3.2.0
-lxml==4.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
-markdown==3.3.6; python_version >= '3.6'
-matplotlib-inline==0.1.3; python_version >= '3.5'
-numpy==1.22.3; python_version >= '3.8'
-parse==1.19.0
-parsley==1.3
-parso==0.8.3; python_version >= '3.6'
-pexpect==4.8.0; sys_platform != 'win32'
-pickleshare==0.7.5
-prompt-toolkit==3.0.28; python_full_version >= '3.6.2'
-psycopg2==2.9.3; python_version >= '3.6'
-ptyprocess==0.7.0
-pure-eval==0.2.2
-pydantic==1.9.0
-pyee==8.2.2
-pyfaidx==0.6.4
-pygments==2.11.2; python_version >= '3.5'
-pyppeteer==1.0.2; python_version >= '3.7' and python_version < '4.0'
-pyquery==1.4.3
-pyrsistent==0.18.1; python_version >= '3.7'
-pysam==0.18.0
-python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
-python-jsonschema-objects==0.4.1
-pyyaml==6.0; python_version >= '3.6'
-requests-html==0.10.0; python_version >= '3.6'
-requests==2.27.1
-s3transfer==0.5.2; python_version >= '3.6'
-setuptools==61.0.0; python_version >= '3.7'
-simplejson==3.17.6; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
-six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
-sniffio==1.2.0; python_version >= '3.5'
-soupsieve==2.3.1; python_version >= '3.6'
-sqlparse==0.4.2; python_version >= '3.5'
-stack-data==0.2.0
-starlette==0.17.1; python_version >= '3.6'
-tabulate==0.8.9
-tqdm==4.63.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
-traitlets==5.1.1; python_version >= '3.7'
-typing-extensions==4.1.1; python_version >= '3.6'
-urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
-uvicorn==0.17.6
-w3lib==1.22.0
-wcwidth==0.2.5
-websockets==10.2; python_version >= '3.7'
-yoyo-migrations==7.3.2
-zipp==3.7.0; python_version >= '3.7'
+anyio==3.6.1 ; python_full_version >= '3.6.2'
+boto3==1.24.92
+botocore==1.27.92 ; python_version >= '3.7'
+click==8.1.3
+fastapi==0.85.1
+ga4gh.vrsatile.pydantic==0.0.11
+h11==0.14.0 ; python_version >= '3.7'
+idna==3.4 ; python_version >= '3.5'
+jmespath==1.0.1 ; python_version >= '3.7'
+pydantic==1.10.2
+python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+s3transfer==0.6.0 ; python_version >= '3.7'
+six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+sniffio==1.3.0 ; python_version >= '3.7'
+starlette==0.20.4 ; python_version >= '3.7'
+typing-extensions==4.4.0 ; python_version >= '3.7'
+urllib3==1.26.12 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
+uvicorn==0.18.3
