+++ setup.cfg
@@ -29,18 +29,19 @@ install_requires =
     uvicorn
     click
     boto3
-    beautifulsoup4
-    gffutils
-    requests
-    biocommons.seqrepo
-    ga4gh.vrs[extras] >= 0.7.5.dev1
-    ga4gh.vrsatile.pydantic >= 0.0.10
-
+    ga4gh.vrsatile.pydantic == 0.0.11
 
 tests_require =
     pytest
     pytest-cov
     mock
+
+[options.extras_require]
+dev =
+    gffutils
+    biocommons.seqrepo
+    ga4gh.vrs[extras] == 0.8.0dev0
+    psycopg2-binary
 
 [tool:pytest]
 addopts = --ignore setup.py --ignore codebuild/ --ignore dynamodb_local_latest/ --doctest-modules --cov-report term-missing --cov .
