+++ tests/unit/test_database_and_etl.py
@@ -91,7 +91,7 @@ def test_ensembl_etl(test_get_seqrepo, processed_ids, dynamodb, etl_data_path,
         shutil.rmtree(e.src_data_dir)
 
         e._sequence_location.get_aliases = _get_aliases
-        e._data_src = etl_data_path / 'ensembl_106.gff3'
+        e._data_src = etl_data_path / 'ensembl_107.gff3'
         e._transform_data()
         e._add_meta()
         processed_ids += e._processed_ids
